Improve state transitions in blob_file_garbage_collector (WIP)
